🔨 refactor: fallback ui를 prop으로 받기
